requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==2.2.10
django-extensions==2.2.7
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.0
